requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==2.3.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-api-logger==1.1.11
drf-writable-nested==0.7.0
drf-yasg==1.21.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pdf2image==1.16.3
Pillow==9.5.0
promise==2.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyocr==0.8.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==3.0.1
pytesseract==0.3.10
pytz==2022.2
razorpay==1.3.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
text-unidecode==1.3
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
webencodings==0.5.1
whitenoise==6.2.0
zope.interface==5.4.0
